// src/auto-gen/enums/attachment-type.enum.ts
export enum AttachmentTypeEnum {
    ATTACHMENT_TYPE_ENUM_UNSPECIFIED = 0,	//Unspecified
    ATTACHMENT_TYPE_ENUM_PHOTO = 1,	//Photo
    ATTACHMENT_TYPE_ENUM_VOICE_MESSAGE = 2,	//Voice message
    ATTACHMENT_TYPE_ENUM_VIDEO_MESSAGE = 3,	//video message
    ATTACHMENT_TYPE_ENUM_AUDIO = 4,	//Audio message
    ATTACHMENT_TYPE_ENUM_VIDEO = 5,	//Video message
    ATTACHMENT_TYPE_ENUM_LINKS = 6,//The link message
    ATTACHMENT_TYPE_ENUM_STICKER = 7,	//The sticker message
    ATTACHMENT_TYPE_ENUM_MEDIA = 8,	//Include PHOTO and VIDEO
    ATTACHMENT_TYPE_ENUM_MENTION = 9,	//Include mention
    ATTACHMENT_TYPE_ENUM_LOCATION = 10,	//Include location
    ATTACHMENT_TYPE_ENUM_FILE = 11,//Include file
}